@@ -15,8 +15,8 @@ class NotesPage extends StatefulWidget {
 }
 
 class _NotesPageState extends State<NotesPage> {
-  List<Note> notes;
-  List<Note> importantNotes;
+  late List<Note> notes;
+  late List<Note> importantNotes;
   bool isLoading = false;
   bool isGrid = true;
 
@@ -29,7 +29,7 @@ class _NotesPageState extends State<NotesPage> {
       LifecycleEventHandler(
         resumeCallBack: () async => setState(() {
           refreshNotes();
-        }),
+        }), suspendingCallBack: () async {  },
       ),
     );
   }
@@ -117,7 +117,7 @@ class _NotesPageState extends State<NotesPage> {
                   return GestureDetector(
                     onTap: () async {
                       await Navigator.of(context).push(MaterialPageRoute(
-                        builder: (context) => NoteDetailPage(noteId: note.id),
+                        builder: (context) => NoteDetailPage(noteId: note.id??-99),
                       ));
                       refreshNotes();
                     },
@@ -141,7 +141,7 @@ class _NotesPageState extends State<NotesPage> {
                   return GestureDetector(
                     onTap: () async {
                       await Navigator.of(context).push(MaterialPageRoute(
-                        builder: (context) => NoteDetailPage(noteId: note.id),
+                        builder: (context) => NoteDetailPage(noteId: note.id??-99),
                       ));
                       refreshNotes();
                     },
@@ -164,7 +164,7 @@ class _NotesPageState extends State<NotesPage> {
                     onTap: () async {
                       await Navigator.of(context).push(
                         MaterialPageRoute(
-                          builder: (context) => NoteDetailPage(noteId: note.id),
+                          builder: (context) => NoteDetailPage(noteId: note.id??-99),
                         ),
                       );
                       refreshNotes();
@@ -179,7 +179,7 @@ class _NotesPageState extends State<NotesPage> {
                     margin: EdgeInsets.symmetric(vertical: 10),
                     child: buildGeneralPadding(),
                   )
-                : "",
+                : SizedBox(),
             Expanded(
               child: ListView.builder(
                 shrinkWrap: true,
@@ -190,7 +190,7 @@ class _NotesPageState extends State<NotesPage> {
                     onTap: () async {
                       await Navigator.of(context).push(
                         MaterialPageRoute(
-                          builder: (context) => NoteDetailPage(noteId: note.id),
+                          builder: (context) => NoteDetailPage(noteId: note.id??-99),
                         ),
                       );
                       refreshNotes();
